Whackamole fixes for disappearing service issues
Once the change was in place to ensure that the service no longer disappears
after being added, all sorts of other issues started to occur:

- Hitting the "edit" button without first expanding the service would make the
  service disappear.
- Selecting a "cancel" button when editing a new service would... make the
  service disappear.
- Switching between tabs when in edit mode on a new service would... make the
  service disappear.

This patch introduces "ag-toggle-include", which will toggle between two
templates based on a condition you provide. Between this and modifying the
edit/cancel buttons to use a scoped function that would change a scoped variable
and set the state without notifications, clicking the edit button no longer
results in a disappearing service. The agTabs and agCancelEdit directives were
altered; agCancelEdit forces a reload, and agTabs now does not force
notification. These solve the remaining issues.