[#166] Ensure entity_id and hydrator_name are synced across config segments
Adds an optional argument, `$isAFetchOperation` to
`RestServiceModel::fetch()`; the value is passed to the `fetch` event to
allow listeners to branch return values based on whether the value is
for a fetch, or an update operation. This ensures that we have the
`resource_name` in DB-Connected rest service operations, which also
ensures that we write to the correct db-connected configuration segment;
previously, because `resource_name` was empty, we would write to an
empty key.