Ensure the UI module is loaded
One issue with moving the admin UI into its own module is that it adds new setup
to the skeleton application:

- You need to add the UI module to the composer.json (potentially mitigated by
  pushing it as a dependency of the admin module).
- You need to add the UI module to the `config/development.config.php.dist` file
  (and potentially the file w/o the dist suffix if already in dev mode).

In order to prevent the necessity of these steps, this patch adds an `init()`
method to the admin module that checks to see if the UI module is already
loaded, and, if not, asks the module manager to load it immediately. This
ensures the config is present and merged by the time you get to the MERGE_CONFIG
event, and thus that the whole thing works seamlessly without needing to change
existing code. Also, the ModuleManager is smart enough that if the module is
encountered later, it will not attempt to load it again.