[#226] Update getServiceType() to reflect changes to service names
Originally, this looked for just namespaced segments (`Rest` and `Rpc`).
However, with the changes made for 1.0.4 to remove encoded slashes from
service names, this breaks link injection, as it could not identify a
REST service by the namespace, and thus always identified services as
RPC. The change introduced will now also detect when the service name
contains dash-separated namespaces (`-Rest-`).