Better validation for content negotiation rules
InputFilters require named inputs. Previously, the CN input filter was
attempting to allow setting the following data structure:

```php
[
    'view model class name' => [ /* array of mediatypes */ ],
]
```

This was failing, because we did not have inputs named after the view
model class names.

The changes in this patch push the view-model/mediatype pairs under a
"selectors" key, allowing them to be validated correctly. It does this
by moving the validation logic to a validator, and then creating an
input that composes that validator. The content_name validation was also
moved into an IsString validator.