Refactored (extract method), and streamable response
Refactored the `indexAction()` to extract a number of discrete methods,
making the logic flow more prominent (vs the details). This included
creating separate methods for package creation vs retrieval, as well as
separate methods for creating the commandline string to use when
invoking zf-deploy.

Additionally, implemented a streamable response to reduce memory
overhead and ensure we don't hit memory limits easily when returning a
package.