Added code to migrate legacy development mode commands
Now also updates the `public/index.php` to allow using the old style of
invoking development mode:

- `php public/index.php development enable`
- `php public/index.php development disable`

These now proxy to the `vendor/bin/zf-development-mode` tooling.

This is done primarily to ensure that invoking the admin in Zend Studio
continues to work, as Studio invokes development mode prior to and
following launching the admin.